Constrain text height for completion label
For whatever reason, the latest Playnite SDK appears to enable text
wrapping on the completion status, causing the bar to balloon in size
due to the wrapped characters.

This constraints it to be the same as the "Time Played" label - a bit of
a hack but should go OK for now.